server/requirements.txt
Flask==3.0.0
mysql-connector==2.2.9
Flask-Mail==0.9.1
pytest==7.4.3